chore: rename rule to `no-multiple-none-start-events`
