+++ index.js
@@ -13,7 +13,7 @@ const camundaCloud10Rules = withConfig({
   'message-reference': 'error',
   'no-candidate-users': 'error',
   'no-expression': 'error',
-  'no-multiple-start-events': 'error',
+  'no-multiple-none-start-events': 'error',
   'no-task-schedule': 'error',
   'no-template': 'error',
   'no-zeebe-properties': 'error',
@@ -116,7 +116,7 @@ module.exports = {
     'message-reference': './rules/camunda-cloud/message-reference',
     'no-candidate-users': './rules/camunda-cloud/no-candidate-users',
     'no-expression': './rules/camunda-cloud/no-expression',
-    'no-multiple-start-events': './rules/camunda-cloud/no-multiple-start-events',
+    'no-multiple-none-start-events': './rules/camunda-cloud/no-multiple-none-start-events',
     'no-signal-event-sub-process': './rules/camunda-cloud/no-signal-event-sub-process',
     'no-task-schedule': './rules/camunda-cloud/no-task-schedule',
     'no-template': './rules/camunda-cloud/no-template',
