+++ rules/camunda-cloud/no-multiple-none-start-events.js
