+++ test/config/configs.spec.js
@@ -18,7 +18,7 @@ describe('configs', function() {
     'message-reference': [ 'error', { version: '1.0' } ],
     'no-candidate-users': [ 'error', { version: '1.0' } ],
     'no-expression': [ 'error', { version: '1.0' } ],
-    'no-multiple-start-events' : [ 'error', { version: '1.0' } ],
+    'no-multiple-none-start-events' : [ 'error', { version: '1.0' } ],
     'no-task-schedule': [ 'error', { version: '1.0' } ],
     'no-template': [ 'error', { version: '1.0' } ],
     'no-zeebe-properties': [ 'error', { version: '1.0' } ],
@@ -43,7 +43,7 @@ describe('configs', function() {
     'message-reference': [ 'error', { version: '1.1' } ],
     'no-candidate-users': [ 'error', { version: '1.1' } ],
     'no-expression': [ 'error', { version: '1.1' } ],
-    'no-multiple-start-events' : [ 'error', { version: '1.1' } ],
+    'no-multiple-none-start-events' : [ 'error', { version: '1.1' } ],
     'no-task-schedule': [ 'error', { version: '1.1' } ],
     'no-template': [ 'error', { version: '1.1' } ],
     'no-zeebe-properties': [ 'error', { version: '1.1' } ],
@@ -68,7 +68,7 @@ describe('configs', function() {
     'message-reference': [ 'error', { version: '1.2' } ],
     'no-candidate-users': [ 'error', { version: '1.2' } ],
     'no-expression': [ 'error', { version: '1.2' } ],
-    'no-multiple-start-events' : [ 'error', { version: '1.2' } ],
+    'no-multiple-none-start-events' : [ 'error', { version: '1.2' } ],
     'no-task-schedule': [ 'error', { version: '1.2' } ],
     'no-template': [ 'error', { version: '1.2' } ],
     'no-zeebe-properties': [ 'error', { version: '1.2' } ],
@@ -93,7 +93,7 @@ describe('configs', function() {
     'message-reference': [ 'error', { version: '1.3' } ],
     'no-candidate-users': [ 'error', { version: '1.3' } ],
     'no-expression': [ 'error', { version: '1.3' } ],
-    'no-multiple-start-events' : [ 'error', { version: '1.3' } ],
+    'no-multiple-none-start-events' : [ 'error', { version: '1.3' } ],
     'no-task-schedule': [ 'error', { version: '1.3' } ],
     'no-template': [ 'error', { version: '1.3' } ],
     'no-zeebe-properties': [ 'error', { version: '1.3' } ],
@@ -118,7 +118,7 @@ describe('configs', function() {
     'message-reference': [ 'error', { version: '8.0' } ],
     'no-candidate-users': [ 'error', { version: '8.0' } ],
     'no-expression': [ 'error', { version: '8.0' } ],
-    'no-multiple-start-events' : [ 'error', { version: '8.0' } ],
+    'no-multiple-none-start-events' : [ 'error', { version: '8.0' } ],
     'no-task-schedule': [ 'error', { version: '8.0' } ],
     'no-zeebe-properties': [ 'error', { version: '8.0' } ],
     'sequence-flow-condition': [ 'error', { version: '8.0' } ],
@@ -143,7 +143,7 @@ describe('configs', function() {
     'message-reference': [ 'error', { version: '8.1' } ],
     'no-candidate-users': [ 'error', { version: '8.1' } ],
     'no-expression': [ 'error', { version: '8.1' } ],
-    'no-multiple-start-events' : [ 'error', { version: '8.1' } ],
+    'no-multiple-none-start-events' : [ 'error', { version: '8.1' } ],
     'no-task-schedule': [ 'error', { version: '8.1' } ],
     'sequence-flow-condition': [ 'error', { version: '8.1' } ],
     'subscription': [ 'error', { version: '8.1' } ],
@@ -167,7 +167,7 @@ describe('configs', function() {
     'loop-characteristics': [ 'error', { version: '8.2' } ],
     'message-reference': [ 'error', { version: '8.2' } ],
     'no-expression': [ 'error', { version: '8.2' } ],
-    'no-multiple-start-events' : [ 'error', { version: '8.2' } ],
+    'no-multiple-none-start-events' : [ 'error', { version: '8.2' } ],
     'no-signal-event-sub-process': [ 'error', { version: '8.2' } ],
     'sequence-flow-condition': [ 'error', { version: '8.2' } ],
     'subscription': [ 'error', { version: '8.2' } ],
@@ -192,7 +192,7 @@ describe('configs', function() {
     'loop-characteristics': [ 'error', { version: '8.3' } ],
     'message-reference': [ 'error', { version: '8.3' } ],
     'no-expression': [ 'error', { version: '8.3' } ],
-    'no-multiple-start-events' : [ 'error', { version: '8.3' } ],
+    'no-multiple-none-start-events' : [ 'error', { version: '8.3' } ],
     'no-signal-event-sub-process': [ 'error', { version: '8.3' } ],
     'sequence-flow-condition': [ 'error', { version: '8.3' } ],
     'signal-reference': [ 'error', { version: '8.3' } ],
